% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{sdEquationList}
\alias{sdEquationList}
\title{Create a list of equations}
\usage{
sdEquationList(...)
}
\arguments{
\item{...}{One or more equations. The equations must have the format
<equation_name> = <equation>, where equation_name is the name of the 
equation, and equation is the body of the equation. The equations must
written in plain R code.}
}
\value{
A list of strings containing the equations in chracter format.
}
\description{
This function converts equations written in plain R code into a list of 
equations in chracter format that can be recognized by the sdsim functions 
\code{\link{sdOdeModel}} and \code{\link{sdStaticModel}}.
}
\examples{
# Create equations list
equations <- sdEquationsList(
  x = a + b,
  y = c ^ 2
)
}
