% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{sdExcelTemplate}
\alias{sdExcelTemplate}
\title{Create a Scenario EXCEL Template}
\usage{
sdExcelTemplate(file = "Scenario.xlsx", colWidth = c(10, 10, 10, 30,
  10))
}
\arguments{
\item{file}{A string with the name of the excel file. The file extension
must be included in the file name, e.g. '.xlsx'.}

\item{colWidth}{The Excel columns width.}
}
\description{
Create a sdsim scenario excel template to be filled by the user and loaded
with the \code{\link{sdLoadScenario}} function. The empty sheets that will
not be used should be deleted before loading the scenario to prevent 
warnings.
}
\examples{
# create a scenario EXCEL template
sdExcelTemplate("test.xlsx")

# edit the created EXCEL file and load it
scenTest <- sdLoadScenario("test.xlsx")
print(scenTest)
}
