% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{sdFlow}
\alias{sdFlow}
\title{Create flow object}
\usage{
sdFlow(flows = NULL, flowRate = NULL, stocks = NULL,
  boundaries = c("boundary"))
}
\arguments{
\item{flows}{}

\item{boundaries}{}
}
\value{

}
\description{
This function converts written text to a list of variables in character 
format that contains the flow model. It also guarantees that the arguments to 
are adequated to be used by other sdsim functions.
}
\examples{
flows <- sdFlow(
  connections = c("birth -> prey", 
                  "prey -> death",
                  "birth -> predator",
                  "predator -> death"),
  flow_rate = c(
    sdsim::sdEquationList(
      st$prey * par$a,
      par$b * st$prey * st$predator,
      par$delta * st$prey * st$predator,
      par$gamma * st$predator
    )
  ),
  boundaries = c("birth", "death"),
  st = c("prey", "predator")
)
scen <- sdScenario(
  id = "test",
  state = list(prey = 10,
               predator = 10),
  parameter = list(a = 2/3,
                   b = 4/3,
                   delta = 1,
                   gamma = 1),
  times = list(from = 0, to = 100, by = 0.1),
  method = "lsoda"
)

model <- sdOdeModel("test",
                    DifferentialEquations = flows,
                    defaultScenario = scen)
out <- sdSimulate(model)
plot(out)
}
