% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{sdFunction}
\alias{sdFunction}
\title{Create function object}
\usage{
sdFunction(func = NULL)
}
\arguments{
\item{func}{}
}
\value{

}
\description{
This function stores the ode function which must have the following format:
function(t, st, ct, par, inp, sw, aux)
}
\examples{
func <- sdsim::sdFunction(
function(t, st, ct, par, inp, sw, aux) {
dPrey <- par$alpha * st$Y -
      par$beta * st$Y * st$R
dPredator <- par$delta * st$Y * st$R -
      par$gamma * st$R
    return(list(c(dPrey, dPredator)))
  }
)
}
