% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{sdLoadModel}
\alias{sdLoadModel}
\title{Loads a Model from a XML File}
\usage{
sdLoadModel(file, repository = F, timeSeriesDirectory = "")
}
\arguments{
\item{file}{The XML file name with the model specification. If
repository = TRUE, \code{file} is the name of a model from the sdsim 
repository.}

\item{repository}{logical. If TRUE will use the \code{file} name to load a 
model from the sdsim package repository. See \code{\link{sdRepository}} for the
complete list of valid names for loading existing models. Default is FALSE.}

\item{timeSeriesDirectory}{The directory where time series inputs are stored 
(when passing the time series inputs via external files).}
}
\value{
A \code{\link{sdOdeModelClass}}, a \code{\link{sdCoupledModelClass}} or 
a \code{\link{sdStaticModelClass}} object.
}
\description{
A factory function that creates \code{\link{sdOdeModelClass}}, 
\code{\link{sdCoupledModelClass}} or \code{\link{sdStaticModelClass}} object 
from a XML file or from the package 
repository (see the available models in the \code{\link{sdRepository}} 
documentation).
}
\details{
The XML file must be created with the models object method '$saveXml' to 
garantee that it will be correctly loaded and validated (by checking the 
sdsim prefix).
}
\examples{
# Load the Bouncing Ball model from the sdsim repository
bb <- sdLoadModel(file = "BouncingBall", repository = TRUE)

# simulate the model and plot the results
outbb <- sdSimulate(model = bb)
outbb$plot("height speed", multipleYAxis = TRUE, units = TRUE)

# save the model to a XML file and reload it
bb$saveXml(file = "bb.xml")
bb <- sdLoadModel(file = "bb.xml") 
}
