% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{sdRunApp}
\alias{sdRunApp}
\title{The sdsim Shiny Server Interface}
\usage{
sdRunApp(launch.browser = T, port = getOption("shiny.port"),
  host = getOption("shiny.host", "127.0.0.1"))
}
\arguments{
\item{launch.browser}{If true, the system's default web browser will be 
launched automatically after the app is started. Defaults to true in 
interactive sessions only. The value of this parameter can also be a 
function to call with the application's URL.}

\item{port}{The TCP port that the application should listen on. If the port 
is not specified, and the shiny.port option is set (with options(shiny.port = 
XX)), then that port will be used. Otherwise, use a random port.}

\item{host}{The IPv4 address that the application should listen on. Defaults 
to the shiny.host option, if set, or "127.0.0.1" if not. See 
\code{\link[shiny]{runApp}} Details.}
}
\description{
Runs the sdsim Shiny server graphical user interface. Open the given http in 
your browser to access the server interface. 
Uses the \code{\link[shiny]{runApp}} function to launch the sdsim 
application.
}
\examples{
## Open the sdsim user interface using the given http
#> sdRunApp()
}
